+++ Raytrace/DirectSampler.cs
@@ -6,8 +6,8 @@ namespace RayTrace
     {
         public override Vector3 color(int x, int y, int width, int height, Camera camera, RayTracer rayTracer, ICollection<Hitable> hitables)
         {
-            var u = (float)x / width;
-            var v = (float)y / height;
+            var u = (double)x / width;
+            var v = (double)y / height;
 
             var r = camera.GetRay(u, v);
             return rayTracer.color(r, hitables);
