+++ Raytrace/Lens.cs
@@ -2,7 +2,7 @@
 {
     public abstract class Lens
     {
-        public abstract Ray GetRay(Camera camera, float u, float v);
+        public abstract Ray GetRay(Camera camera, double u, double v);
     }
