+++ Raytrace/Metal.cs
@@ -3,7 +3,7 @@
     public class Metal : Material
     {
         public Vector3 Albedo { get; set; }
-        public float Fuzz { get; set; }
+        public double Fuzz { get; set; }
        
 
         public override bool Scatter(Ray r, HitRecord record, out Vector3 attenuation, out Ray scattered)
